Interface scorecard::woebin for partykiy::ctree

Interface scorecard::woebin for partykiy::ctree

Usage

woebin_ctree(
  y,
  x,
  namevar = "variable",
  count_distr_limit = 0.05,
  control = partykit::ctree_control()
)

Arguments

y

A vector of response. Usually 0-1

x

A predictive variable

namevar

a character element

count_distr_limit

The minimum count distribution percentage. Accepted range: 0.01-0.2; Defaults to 0.05.

control

a ctree::ctree_control list element

Examples

data(germancredit, package = "scorecard")

y <- germancredit$creditability

x <- germancredit$duration.in.month

woebin_ctree(y, x, "duration", count_distr_limit = 0.05)
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#>    variable       bin count count_distr neg pos   posprob        woe
#> 1: duration [-Inf,11)   171       0.171 144  27 0.1578947 -0.8266786
#> 2: duration   [11,33)   656       0.656 466 190 0.2896341 -0.0498637
#> 3: duration [33, Inf)   173       0.173  90  83 0.4797688  0.7663288
#>         bin_iv  total_iv breaks is_special_values
#> 1: 0.095658521 0.2107628     11             FALSE
#> 2: 0.001614634 0.2107628     33             FALSE
#> 3: 0.113489646 0.2107628    Inf             FALSE

woebin_ctree(y, x, "duration", count_distr_limit = 0.2)
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#>    variable       bin count count_distr neg pos   posprob        woe
#> 1: duration [-Inf,15)   367       0.367 290  77 0.2098093 -0.4787776
#> 2: duration   [15,28)   417       0.417 291 126 0.3021583  0.0102565
#> 3: duration [28, Inf)   216       0.216 119  97 0.4490741  0.6428853
#>          bin_iv  total_iv breaks is_special_values
#> 1: 7.546448e-02 0.1740842     15             FALSE
#> 2: 4.395643e-05 0.1740842     28             FALSE
#> 3: 9.857575e-02 0.1740842    Inf             FALSE

woebin_ctree(
  y,
  x,
  "duration",
  count_distr_limit = 0.05,
  control = partykit::ctree_control(alpha = 0.5)
)
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#>    variable       bin count count_distr neg pos   posprob         woe
#> 1: duration  [-Inf,8)    87       0.087  78   9 0.1034483 -1.31548096
#> 2: duration    [8,11)    84       0.084  66  18 0.2142857 -0.45527969
#> 3: duration   [11,12)     9       0.009   9   0 0.0000000 -1.36327162
#> 4: duration   [12,15)   187       0.187 137  50 0.2673797 -0.16395463
#> 5: duration   [15,33)   460       0.460 320 140 0.3043478  0.01732472
#> 6: duration   [33,42)    92       0.092  52  40 0.4347826  0.58163903
#> 7: duration [42, Inf)    81       0.081  38  43 0.5308642  0.96761725
#>         bin_iv  total_iv breaks is_special_values
#> 1: 0.107247539 0.2607787      8             FALSE
#> 2: 0.015699438 0.2607787     11             FALSE
#> 3: 0.013043779 0.2607787     12             FALSE
#> 4: 0.004852370 0.2607787     15             FALSE
#> 5: 0.000138405 0.2607787     33             FALSE
#> 6: 0.034089320 0.2607787     42             FALSE
#> 7: 0.085707835 0.2607787    Inf             FALSE

x[sample(c(TRUE, FALSE), size = length(x), prob = c(1, 99), replace = TRUE)] <- NA
woebin_ctree(y, x, "duration")
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#>    variable       bin count count_distr neg pos   posprob        woe     bin_iv
#> 1: duration   missing    10       0.010   7   3 0.3000000  0.0000000 0.00000000
#> 2: duration [-Inf,15)   367       0.367 290  77 0.2098093 -0.4787776 0.07546448
#> 3: duration   [15,33)   450       0.450 313 137 0.3044444  0.0210756 0.00020072
#> 4: duration [33, Inf)   173       0.173  90  83 0.4797688  0.7663288 0.11348965
#>     total_iv  breaks is_special_values
#> 1: 0.1891548 missing              TRUE
#> 2: 0.1891548      15             FALSE
#> 3: 0.1891548      33             FALSE
#> 4: 0.1891548     Inf             FALSE

x <- germancredit$purpose
woebin_ctree(y, x, "purpose")
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#> Warning: The categorical values (`radio/television, education, furniture/equipment, car (new), car (used), business, domestic appliances, repairs, others, retraining`) are not specified in `breaks_list` for the column `purpose`.
#> Warning: The following columns are removed from binning results due to errors:
#> purpose
#>    variable
#> 1:  purpose
#> 2:  purpose
#>                                                                                                bin
#> 1: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 2:                                                      radio/television%,%car (used)%,%retraining
#>    count count_distr neg pos   posprob        woe     bin_iv total_iv
#> 1:   608       0.608 388 220 0.3618421  0.2799201 0.05011902 0.142092
#> 2:   392       0.392 312  80 0.2040816 -0.5136787 0.09197295 0.142092
#>                                                                                             breaks
#> 1: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 2:                                                      radio/television%,%car (used)%,%retraining
#>    is_special_values
#> 1:             FALSE
#> 2:             FALSE

x[sample(c(TRUE, FALSE), size = length(x), prob = c(1, 99), replace = TRUE)] <- NA
woebin_ctree(y, x, "purpose_with_na")
#> [INFO] creating woe binning ... 
#> [INFO] The option bin_close_right was set to FALSE. Resetting by options(scorecard.bin_close_right = FALSE).
#> Warning: The categorical values (`radio/television, education, furniture/equipment, car (new), car (used), business, domestic appliances, repairs, others, retraining`) are not specified in `breaks_list` for the column `purpose_with_na`.
#> Warning: The following columns are removed from binning results due to errors:
#> purpose_with_na
#>           variable
#> 1: purpose_with_na
#> 2: purpose_with_na
#> 3: purpose_with_na
#>                                                                                                bin
#> 1:                                                                                         missing
#> 2: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 3:                                                      radio/television%,%car (used)%,%retraining
#>    count count_distr neg pos   posprob        woe     bin_iv  total_iv
#> 1:    15       0.015   8   7 0.4666667  0.7137665 0.00849722 0.1459771
#> 2:   600       0.600 385 215 0.3583333  0.2646926 0.04411543 0.1459771
#> 3:   385       0.385 307  78 0.2025974 -0.5228411 0.09336448 0.1459771
#>                                                                                             breaks
#> 1:                                                                                         missing
#> 2: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 3:                                                      radio/television%,%car (used)%,%retraining
#>    is_special_values
#> 1:              TRUE
#> 2:             FALSE
#> 3:             FALSE