Get predictive indicator for partial models given a model

Get predictive indicator for partial models given a model

model_partials(model, newdata = NULL)

Examples

library(scorecard)
data("germancredit")

dt    <-  var_filter(germancredit, "creditability")

#> [INFO] filtering variables ... 

bins  <-  woebin(dt, "creditability")

#> [INFO] creating woe binning ... 

dtwoe <-  woebin_ply(dt, bins)

#> [INFO] converting into woe values ... 

model <- glm(creditability ~ ., family = binomial(), data = dtwoe)

model_partials(model)

#> # A tibble: 13 x 5
#>    variable                                                 ks   auc    iv  gini
#>    <fct>                                                 <dbl> <dbl> <dbl> <dbl>
#>  1 status.of.existing.checking.account_woe               0.367 0.690 0.584 0.380
#>  2 duration.in.month_woe                                 0.385 0.748 0.877 0.497
#>  3 credit.history_woe                                    0.432 0.770 1.11  0.541
#>  4 purpose_woe                                           0.440 0.785 1.31  0.570
#>  5 credit.amount_woe                                     0.473 0.797 1.34  0.595
#>  6 savings.account.and.bonds_woe                         0.477 0.804 1.52  0.608
#>  7 present.employment.since_woe                          0.491 0.809 1.46  0.618
#>  8 installment.rate.in.percentage.of.disposable.income_~ 0.497 0.813 1.53  0.627
#>  9 other.debtors.or.guarantors_woe                       0.507 0.816 1.61  0.633
#> 10 property_woe                                          0.499 0.818 1.57  0.635
#> 11 age.in.years_woe                                      0.519 0.822 1.72  0.645
#> 12 other.installment.plans_woe                           0.532 0.825 1.71  0.651
#> 13 housing_woe                                           0.528 0.827 1.73  0.654