Get predictive indicator for partial models given a model

Get predictive indicator for partial models given a model

Usage

model_partials(model, newdata = NULL, verbose = TRUE)

Arguments

model

model

newdata

Optional data frame

verbose

verbose

Examples

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = head(credit_woe, 10000))
m <- featsel_stepforward(m)
#> Start:  AIC=9982.26
#> bad ~ 1
#> 
#>                           Df Deviance    AIC
#> + age_woe                  1   9660.2 9664.2
#> + months_in_the_job_woe    1   9838.6 9842.6
#> + marital_status_woe       1   9846.1 9850.1
#> + flag_res_phone_woe       1   9857.1 9861.1
#> + profession_code_woe      1   9900.2 9904.2
#> + personal_net_income_woe  1   9924.3 9928.3
#> + area_code_res_phone_woe  1   9924.4 9928.4
#> + sex_woe                  1   9938.7 9942.7
#> + months_in_residence_woe  1   9947.3 9951.3
#> + quant_add_cards_woe      1   9947.6 9951.6
#> + payment_day_woe          1   9952.2 9956.2
#> + residence_type_woe       1   9971.2 9975.2
#> + id_client_woe            1   9971.7 9975.7
#> <none>                         9980.3 9982.3
#> 
#> Step:  AIC=9664.17
#> bad ~ age_woe
#> 
#>                           Df Deviance    AIC
#> + flag_res_phone_woe       1   9556.6 9562.6
#> + area_code_res_phone_woe  1   9607.0 9613.0
#> + months_in_the_job_woe    1   9621.0 9627.0
#> + payment_day_woe          1   9637.6 9643.6
#> + sex_woe                  1   9638.6 9644.6
#> + marital_status_woe       1   9646.9 9652.9
#> + profession_code_woe      1   9647.7 9653.7
#> + id_client_woe            1   9652.6 9658.6
#> + personal_net_income_woe  1   9652.8 9658.8
#> + quant_add_cards_woe      1   9653.3 9659.3
#> <none>                         9660.2 9664.2
#> + months_in_residence_woe  1   9659.6 9665.6
#> + residence_type_woe       1   9660.2 9666.2
#> - age_woe                  1   9980.3 9982.3
#> 
#> Step:  AIC=9562.61
#> bad ~ age_woe + flag_res_phone_woe
#> 
#>                           Df Deviance    AIC
#> + months_in_the_job_woe    1   9520.3 9528.3
#> + payment_day_woe          1   9537.8 9545.8
#> + sex_woe                  1   9540.4 9548.4
#> + marital_status_woe       1   9545.8 9553.8
#> + profession_code_woe      1   9546.0 9554.0
#> + area_code_res_phone_woe  1   9547.1 9555.1
#> + id_client_woe            1   9549.4 9557.4
#> + quant_add_cards_woe      1   9553.4 9561.4
#> <none>                         9556.6 9562.6
#> + personal_net_income_woe  1   9554.8 9562.8
#> + months_in_residence_woe  1   9556.2 9564.2
#> + residence_type_woe       1   9556.6 9564.6
#> - flag_res_phone_woe       1   9660.2 9664.2
#> - age_woe                  1   9857.1 9861.1
#> 
#> Step:  AIC=9528.27
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe
#> 
#>                           Df Deviance    AIC
#> + payment_day_woe          1   9499.2 9509.2
#> + sex_woe                  1   9501.2 9511.2
#> + profession_code_woe      1   9507.6 9517.6
#> + area_code_res_phone_woe  1   9510.1 9520.1
#> + marital_status_woe       1   9511.8 9521.8
#> + id_client_woe            1   9513.0 9523.0
#> <none>                         9520.3 9528.3
#> + quant_add_cards_woe      1   9519.1 9529.1
#> + months_in_residence_woe  1   9520.1 9530.1
#> + residence_type_woe       1   9520.2 9530.2
#> + personal_net_income_woe  1   9520.3 9530.3
#> - months_in_the_job_woe    1   9556.6 9562.6
#> - flag_res_phone_woe       1   9621.0 9627.0
#> - age_woe                  1   9724.7 9730.7
#> 
#> Step:  AIC=9509.15
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe
#> 
#>                           Df Deviance    AIC
#> + sex_woe                  1   9479.3 9491.3
#> + profession_code_woe      1   9487.9 9499.9
#> + area_code_res_phone_woe  1   9489.4 9501.4
#> + marital_status_woe       1   9490.8 9502.8
#> + id_client_woe            1   9491.4 9503.4
#> <none>                         9499.2 9509.2
#> + quant_add_cards_woe      1   9498.2 9510.2
#> + months_in_residence_woe  1   9499.0 9511.0
#> + residence_type_woe       1   9499.1 9511.1
#> + personal_net_income_woe  1   9499.1 9511.1
#> - payment_day_woe          1   9520.3 9528.3
#> - months_in_the_job_woe    1   9537.8 9545.8
#> - flag_res_phone_woe       1   9595.8 9603.8
#> - age_woe                  1   9697.6 9705.6
#> 
#> Step:  AIC=9491.25
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe
#> 
#>                           Df Deviance    AIC
#> + profession_code_woe      1   9466.9 9480.9
#> + area_code_res_phone_woe  1   9469.2 9483.2
#> + marital_status_woe       1   9470.0 9484.0
#> + id_client_woe            1   9471.6 9485.6
#> + quant_add_cards_woe      1   9475.1 9489.1
#> <none>                         9479.3 9491.3
#> + personal_net_income_woe  1   9478.9 9492.9
#> + months_in_residence_woe  1   9479.1 9493.1
#> + residence_type_woe       1   9479.2 9493.2
#> - sex_woe                  1   9499.2 9509.2
#> - payment_day_woe          1   9501.2 9511.2
#> - months_in_the_job_woe    1   9521.0 9531.0
#> - flag_res_phone_woe       1   9569.9 9579.9
#> - age_woe                  1   9661.1 9671.1
#> 
#> Step:  AIC=9480.9
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe + profession_code_woe
#> 
#>                           Df Deviance    AIC
#> + marital_status_woe       1   9457.4 9473.4
#> + area_code_res_phone_woe  1   9457.6 9473.6
#> + id_client_woe            1   9459.5 9475.5
#> + quant_add_cards_woe      1   9463.5 9479.5
#> <none>                         9466.9 9480.9
#> + personal_net_income_woe  1   9466.7 9482.7
#> + months_in_residence_woe  1   9466.8 9482.8
#> + residence_type_woe       1   9466.9 9482.9
#> - profession_code_woe      1   9479.3 9491.3
#> - payment_day_woe          1   9487.3 9499.3
#> - sex_woe                  1   9487.9 9499.9
#> - months_in_the_job_woe    1   9511.0 9523.0
#> - flag_res_phone_woe       1   9555.6 9567.6
#> - age_woe                  1   9602.6 9614.6
#> 
#> Step:  AIC=9473.43
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe + profession_code_woe + marital_status_woe
#> 
#>                           Df Deviance    AIC
#> + area_code_res_phone_woe  1   9449.0 9467.0
#> + id_client_woe            1   9450.2 9468.2
#> + quant_add_cards_woe      1   9455.3 9473.3
#> <none>                         9457.4 9473.4
#> + residence_type_woe       1   9457.2 9475.2
#> + months_in_residence_woe  1   9457.2 9475.2
#> + personal_net_income_woe  1   9457.4 9475.4
#> - marital_status_woe       1   9466.9 9480.9
#> - profession_code_woe      1   9470.0 9484.0
#> - payment_day_woe          1   9477.8 9491.8
#> - sex_woe                  1   9479.3 9493.3
#> - months_in_the_job_woe    1   9498.8 9512.8
#> - age_woe                  1   9543.5 9557.5
#> - flag_res_phone_woe       1   9543.9 9557.9
#> 
#> Step:  AIC=9466.99
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe + profession_code_woe + marital_status_woe + 
#>     area_code_res_phone_woe
#> 
#>                           Df Deviance    AIC
#> + id_client_woe            1   9442.1 9462.1
#> + quant_add_cards_woe      1   9446.8 9466.8
#> <none>                         9449.0 9467.0
#> + months_in_residence_woe  1   9448.7 9468.7
#> + residence_type_woe       1   9448.8 9468.8
#> + personal_net_income_woe  1   9449.0 9469.0
#> - area_code_res_phone_woe  1   9457.4 9473.4
#> - marital_status_woe       1   9457.6 9473.6
#> - profession_code_woe      1   9460.8 9476.8
#> - payment_day_woe          1   9469.0 9485.0
#> - sex_woe                  1   9471.2 9487.2
#> - months_in_the_job_woe    1   9491.1 9507.1
#> - flag_res_phone_woe       1   9498.4 9514.4
#> - age_woe                  1   9536.7 9552.7
#> 
#> Step:  AIC=9462.14
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe + profession_code_woe + marital_status_woe + 
#>     area_code_res_phone_woe + id_client_woe
#> 
#>                           Df Deviance    AIC
#> + quant_add_cards_woe      1   9440.0 9462.0
#> <none>                         9442.1 9462.1
#> + months_in_residence_woe  1   9441.9 9463.9
#> + residence_type_woe       1   9442.0 9464.0
#> + personal_net_income_woe  1   9442.1 9464.1
#> - id_client_woe            1   9449.0 9467.0
#> - area_code_res_phone_woe  1   9450.2 9468.2
#> - marital_status_woe       1   9450.6 9468.6
#> - profession_code_woe      1   9453.8 9471.8
#> - payment_day_woe          1   9462.7 9480.7
#> - sex_woe                  1   9464.2 9482.2
#> - months_in_the_job_woe    1   9484.3 9502.3
#> - flag_res_phone_woe       1   9491.7 9509.7
#> - age_woe                  1   9529.7 9547.7
#> 
#> Step:  AIC=9461.96
#> bad ~ age_woe + flag_res_phone_woe + months_in_the_job_woe + 
#>     payment_day_woe + sex_woe + profession_code_woe + marital_status_woe + 
#>     area_code_res_phone_woe + id_client_woe + quant_add_cards_woe
#> 
#>                           Df Deviance    AIC
#> <none>                         9440.0 9462.0
#> - quant_add_cards_woe      1   9442.1 9462.1
#> + months_in_residence_woe  1   9439.7 9463.7
#> + residence_type_woe       1   9439.8 9463.8
#> + personal_net_income_woe  1   9439.9 9463.9
#> - id_client_woe            1   9446.8 9466.8
#> - marital_status_woe       1   9447.2 9467.2
#> - area_code_res_phone_woe  1   9448.1 9468.1
#> - profession_code_woe      1   9451.0 9471.0
#> - payment_day_woe          1   9460.3 9480.3
#> - sex_woe                  1   9464.0 9484.0
#> - months_in_the_job_woe    1   9479.7 9499.7
#> - flag_res_phone_woe       1   9487.4 9507.4
#> - age_woe                  1   9525.0 9545.0

model_partials(m)
#> Fitting and evaluating model with 1 variables: ~Fitting and evaluating model with 1 variables: badFitting and evaluating model with 1 variables: age_woe
#> Fitting and evaluating model with 2 variables: ~Fitting and evaluating model with 2 variables: badFitting and evaluating model with 2 variables: age_woe + flag_res_phone_woe
#> Fitting and evaluating model with 3 variables: ~Fitting and evaluating model with 3 variables: badFitting and evaluating model with 3 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 4 variables: ~Fitting and evaluating model with 4 variables: badFitting and evaluating model with 4 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 5 variables: ~Fitting and evaluating model with 5 variables: badFitting and evaluating model with 5 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 6 variables: ~Fitting and evaluating model with 6 variables: badFitting and evaluating model with 6 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 7 variables: ~Fitting and evaluating model with 7 variables: badFitting and evaluating model with 7 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 8 variables: ~Fitting and evaluating model with 8 variables: badFitting and evaluating model with 8 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 9 variables: ~Fitting and evaluating model with 9 variables: badFitting and evaluating model with 9 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 10 variables: ~Fitting and evaluating model with 10 variables: badFitting and evaluating model with 10 variables: age_woe + flag_res_phone_woe ...
#> # A tibble: 10 × 5
#>    variable                   ks   auc    iv  gini
#>    <fct>                   <dbl> <dbl> <dbl> <dbl>
#>  1 age_woe                 0.204 0.624 0.190 0.248
#>  2 flag_res_phone_woe      0.207 0.641 0.248 0.283
#>  3 months_in_the_job_woe   0.223 0.650 0.276 0.301
#>  4 payment_day_woe         0.238 0.656 0.304 0.312
#>  5 sex_woe                 0.247 0.659 0.309 0.319
#>  6 profession_code_woe     0.244 0.661 0.329 0.321
#>  7 marital_status_woe      0.244 0.662 0.335 0.324
#>  8 area_code_res_phone_woe 0.245 0.664 0.346 0.328
#>  9 id_client_woe           0.250 0.665 0.338 0.330
#> 10 quant_add_cards_woe     0.250 0.665 0.336 0.331

model_partials(m, newdata = tail(credit_woe, 10000))
#> Fitting and evaluating model with 1 variables: ~Fitting and evaluating model with 1 variables: badFitting and evaluating model with 1 variables: age_woe
#> Fitting and evaluating model with 2 variables: ~Fitting and evaluating model with 2 variables: badFitting and evaluating model with 2 variables: age_woe + flag_res_phone_woe
#> Fitting and evaluating model with 3 variables: ~Fitting and evaluating model with 3 variables: badFitting and evaluating model with 3 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 4 variables: ~Fitting and evaluating model with 4 variables: badFitting and evaluating model with 4 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 5 variables: ~Fitting and evaluating model with 5 variables: badFitting and evaluating model with 5 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 6 variables: ~Fitting and evaluating model with 6 variables: badFitting and evaluating model with 6 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 7 variables: ~Fitting and evaluating model with 7 variables: badFitting and evaluating model with 7 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 8 variables: ~Fitting and evaluating model with 8 variables: badFitting and evaluating model with 8 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 9 variables: ~Fitting and evaluating model with 9 variables: badFitting and evaluating model with 9 variables: age_woe + flag_res_phone_woe +...
#> Fitting and evaluating model with 10 variables: ~Fitting and evaluating model with 10 variables: badFitting and evaluating model with 10 variables: age_woe + flag_res_phone_woe ...
#> # A tibble: 20 × 6
#>    variable                sample    ks   auc    iv  gini
#>    <fct>                   <fct>  <dbl> <dbl> <dbl> <dbl>
#>  1 age_woe                 train  0.204 0.624 0.190 0.248
#>  2 age_woe                 test   0.194 0.624 0.213 0.249
#>  3 flag_res_phone_woe      train  0.207 0.641 0.248 0.283
#>  4 flag_res_phone_woe      test   0.208 0.645 0.262 0.289
#>  5 months_in_the_job_woe   train  0.223 0.650 0.276 0.301
#>  6 months_in_the_job_woe   test   0.225 0.654 0.312 0.307
#>  7 payment_day_woe         train  0.238 0.656 0.304 0.312
#>  8 payment_day_woe         test   0.234 0.658 0.335 0.317
#>  9 sex_woe                 train  0.247 0.659 0.309 0.319
#> 10 sex_woe                 test   0.241 0.665 0.344 0.329
#> 11 profession_code_woe     train  0.244 0.661 0.329 0.321
#> 12 profession_code_woe     test   0.249 0.669 0.365 0.337
#> 13 marital_status_woe      train  0.244 0.662 0.335 0.324
#> 14 marital_status_woe      test   0.251 0.671 0.385 0.341
#> 15 area_code_res_phone_woe train  0.245 0.664 0.346 0.328
#> 16 area_code_res_phone_woe test   0.259 0.674 0.397 0.347
#> 17 id_client_woe           train  0.250 0.665 0.338 0.330
#> 18 id_client_woe           test   0.259 0.675 0.405 0.351
#> 19 quant_add_cards_woe     train  0.250 0.665 0.336 0.331
#> 20 quant_add_cards_woe     test   0.259 0.676 0.402 0.352