Interface scorecard::woebin for partykiy::ctree

Interface scorecard::woebin for partykiy::ctree

woebin_ctree(
  y,
  x,
  namevar = "variable",
  count_distr_limit = 0.05,
  control = partykit::ctree_control()
)

Arguments

y
A vector of response. Usually 0-1

x
A predictive variable

namevar
a character element

count_distr_limit
The minimum count distribution percentage. Accepted range: 0.01-0.2; Defaults to 0.05.

control
a ctree::ctree_control list element

Examples

data(germancredit, package = "scorecard")

y <- germancredit$creditability

x <- germancredit$duration.in.month

woebin_ctree(y, x, "duration", count_distr_limit = 0.05)

#> [INFO] creating woe binning ... 

#>    variable         bin count count_distr neg pos posprob woe bin_iv total_iv
#> 1: duration [-Inf, Inf)  1000           1 700 300     0.3   0      0        0
#>    breaks is_special_values
#> 1:    Inf             FALSE

woebin_ctree(y, x, "duration", count_distr_limit = 0.2)

#> [INFO] creating woe binning ... 

#>    variable         bin count count_distr neg pos posprob woe bin_iv total_iv
#> 1: duration [-Inf, Inf)  1000           1 700 300     0.3   0      0        0
#>    breaks is_special_values
#> 1:    Inf             FALSE

woebin_ctree(
  y,
  x,
  "duration",
  count_distr_limit = 0.05,
  control = partykit::ctree_control(alpha = 0.5)
)

#> [INFO] creating woe binning ... 

#>    variable       bin count count_distr neg pos   posprob         woe
#> 1: duration [-Inf,42)   919       0.919 662 257 0.2796518 -0.09889161
#> 2: duration [42, Inf)    81       0.081  38  43 0.5308642  0.97091182
#>         bin_iv   total_iv breaks is_special_values
#> 1: 0.008806062 0.09526345     42             FALSE
#> 2: 0.086457386 0.09526345    Inf             FALSE

x[sample(c(TRUE, FALSE), size = length(x), prob = c(1, 99), replace = TRUE)] <- NA
woebin_ctree(y, x, "duration")

#> [INFO] creating woe binning ... 

#>    variable       bin count count_distr neg pos   posprob        woe
#> 1: duration   missing     9       0.009   6   3 0.3333333  0.1541507
#> 2: duration [-Inf,42)   910       0.910 656 254 0.2791209 -0.1015287
#> 3: duration [42, Inf)    81       0.081  38  43 0.5308642  0.9709118
#>          bin_iv   total_iv  breaks is_special_values
#> 1: 0.0002202153 0.09586353 missing              TRUE
#> 2: 0.0091859265 0.09586353      42             FALSE
#> 3: 0.0864573856 0.09586353     Inf             FALSE

x <- germancredit$purpose
woebin_ctree(y, x, "purpose")

#> [INFO] creating woe binning ... 

#> Warning: The categorical values (`radio/television, education, furniture/equipment, car (new), car (used), business, domestic appliances, repairs, others, retraining`) are not specified in `breaks_list` for the column `purpose`.

#> Warning: The following columns are removed from binning results due to errors:
#> purpose

#>    variable
#> 1:  purpose
#> 2:  purpose
#>                                                                                                bin
#> 1: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 2:                                                      radio/television%,%car (used)%,%retraining
#>    count count_distr neg pos   posprob        woe     bin_iv total_iv
#> 1:   608       0.608 388 220 0.3618421  0.2799201 0.05011902 0.142092
#> 2:   392       0.392 312  80 0.2040816 -0.5136787 0.09197295 0.142092
#>                                                                                             breaks
#> 1: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 2:                                                      radio/television%,%car (used)%,%retraining
#>    is_special_values
#> 1:             FALSE
#> 2:             FALSE

x[sample(c(TRUE, FALSE), size = length(x), prob = c(1, 99), replace = TRUE)] <- NA
woebin_ctree(y, x, "purpose_with_na")

#> [INFO] creating woe binning ... 

#> Warning: The categorical values (`radio/television, education, furniture/equipment, car (new), car (used), business, domestic appliances, repairs, others, retraining`) are not specified in `breaks_list` for the column `purpose_with_na`.

#> Warning: The following columns are removed from binning results due to errors:
#> purpose_with_na

#>           variable
#> 1: purpose_with_na
#> 2: purpose_with_na
#> 3: purpose_with_na
#>                                                                                                bin
#> 1:                                                                                         missing
#> 2: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 3:                                                      radio/television%,%car (used)%,%retraining
#>    count count_distr neg pos   posprob        woe       bin_iv total_iv
#> 1:    18       0.018  12   6 0.3333333  0.1541507 0.0004404305 0.138619
#> 2:   599       0.599 383 216 0.3606010  0.2745413 0.0474564211 0.138619
#> 3:   383       0.383 305  78 0.2036554 -0.5163051 0.0907221800 0.138619
#>                                                                                             breaks
#> 1:                                                                                         missing
#> 2: education%,%furniture/equipment%,%car (new)%,%business%,%domestic appliances%,%repairs%,%others
#> 3:                                                      radio/television%,%car (used)%,%retraining
#>    is_special_values
#> 1:              TRUE
#> 2:             FALSE
#> 3:             FALSE